requirements.txt
aiohttp-cors==0.7.0
appdirs==1.4.4
astor==0.8.1
atomicwrites==1.4.0
bidict==0.23.1
blake3==1.0.1
colorful==0.5.6
compressed-tensors==0.8.1
datasets==3.2.0
deprecation==2.1.0
depyf==0.18.0
diskcache==5.6.3
distlib==0.3.9
duckdb==1.1.3
einops==0.8.0
fastapi==0.115.6
filelock==3.16.1
fsspec==2023.12.2
gguf==0.10.0
grpcio==1.69.0
gurobipy==12.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
interegular==0.3.3
jiter==0.8.2
jsonpointer==2.1
jsonschema==4.23.0
lark==1.2.2
lm-format-enforcer==0.10.9
memray==1.15.0
mistral_common==1.5.1
mkl-service==2.4.0
mmh3==4.1.0
msgspec==0.19.0
multiprocess==0.70.16
nvidia-ml-py==12.560.30
openai==1.57.4
outlines==0.1.11
outlines_core==0.1.26
partial-json-parser==0.2.1.1.post4
progressbar==2.5
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
proto-plus==1.25.0
protobuf==3.20.3
py-spy==0.4.0
pyarrow==18.1.0
pyasn1-modules==0.2.8
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pyiceberg==0.7.1
pyls-spyder==0.4.0
pyparsing==3.2.0
ray==2.40.0
rsa==4.9
s3fs==2023.12.2
safetensors==0.5.0
sentencepiece==0.2.0
setuptools==75.6.0
starlette==0.41.3
strictyaml==1.7.3
sympy==1.13.1
tenacity==8.5.0
textual==1.0.0
tiktoken==0.7.0
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
vllm==0.6.6.post1
watchfiles==1.0.3
websockets==14.1
wheel==0.43.0
xformers==0.0.28.post3
xgrammar==0.1.8
xxhash==3.5.0
